Move MultiBankApplication to root package
