+++ src/main/java/com/multibank/MultiBankApplication.java
@@ -1,4 +1,4 @@
-package com.multibank.application;
+package com.multibank;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
