+++ src/test/java/com/multibank/MultiBankApplicationTests.java
@@ -1,6 +1,6 @@
 package com.multibank;
 
-import com.multibank.application.MultiBankApplication;
+import com.multibank.MultiBankApplication;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
